requirements.txt
streamlit>=1.37
pandas>=2.2
pydantic>=2.7
openai>=1.30
requests>=2.32
chromadb>=0.4
python-dotenv>=1.0
numpy>=1.26
scikit-learn>=1.4
openpyxl>=3.1
pyyaml>=6.0
scipy>=1.11

pytest>=7.0
